requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.7
contourpy==1.3.3
cycler==0.12.1
fake-http-header==0.3.5
fonttools==4.62.1
greenlet==3.3.2
idna==3.11
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pillow==12.2.0
playwright==1.58.0
pyee==13.0.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
seaborn==0.13.2
six==1.17.0
soupsieve==2.8.3
tf-playwright-stealth==1.2.0
typing_extensions==4.15.0
urllib3==2.6.3
